Add thread as a common method.
